+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All Notable changes to `jobs-github` will be documented in this file
 
+## 1.0.0 - 2016-09-07
+
+### Added
+- Support for v2 of jobs-common package
+- New Query and Provider classes
+- E2E and integration tests
+
+### Removed
+- Support for all old `set<Parameter>` methods in the Provider.
+
 ## 0.1.1 - 2015-08-30
 
 ### Added